requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.3
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
beautifulsoup4>=4.10.0
playwright>=1.30.0
pillow>=9.0.0
python-dotenv>=0.19.2
apscheduler>=3.9.1
celery>=5.2.7
redis>=4.3.4
streamlit>=1.12.0
pytest>=7.0.0
httpx>=0.23.0
python-multipart>=0.0.5
tenacity>=8.0.1
alembic>=1.8.1
requests>=2.27.1
openai>=1.0.0
tiktoken>=0.5.0
chromadb>=0.4.0 